Refactor grid system to be embeded in wrapping components & rename wrapping components
